requirements.txt
alembic>=1.13.0
Flask>=3.0.0
Flask-Login>=0.6.0
Flask-SQLAlchemy>=3.1.0
openai>=1.0.0
python-dotenv>=1.0.0
pytest>=7.0.0
pydantic>=2.0.0
